requirements.txt
google-adk==1.4.2
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.98.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-datastore==2.21.0
google-cloud-discoveryengine==0.13.9
google-cloud-language==2.17.1
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.54.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-cloud-translate==3.20.2
google-cloud-videointelligence==2.16.1
google-cloud-vision==3.10.1
google-crc32c==1.7.1
google-genai==1.21.1
gce-v2v @ https://storage.googleapis.com/velostrata-release/gce-v2v/gce-v2v.tar.gz#sha256=1533c75fd582af8102f3c7bb800342521b0964266f723e29ebd5afddc1089c4d
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
sendgrid
keyrings.google-artifactregistry-auth==1.1.2
flask[async]
groq
protobuf==6.31.1
python-dotenv==1.1.0
gunicorn==23.0.0
